DOC: Adds a tract profile plot to the ALS example.
Also, fixes a small bug in the plot_tract_profile function, which assumes that
the inputs are arrays, by referring to the `group_by.shape` attribute.

This is one option for addressing this. Another would be to require this
input to be an array.

I also changed the example to download the data and load it as two separate
steps, because I don't think that the automagical version is as useful for
someone getting started with the software.